+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Timeregistration: Fixed time dialog tracking time starting at midnight
 - Newsletters: fix missing property error when sending newsletter with attachment
+- Smime: Handle two valid certficates for the same sender
 
 08-10-2024: 6.8.76
 - ActiveSync: Fixed fatal error because of breaking change in z-push